@@ -1,14 +1,23 @@
-// swift-tools-version:5.9
+// swift-tools-version:6.0
 import PackageDescription
 
 let package = Package(
     name: "LuxAnalytics",
-    platforms: [.iOS(.v13)],
+    platforms: [.iOS(.v18)],
     products: [
         .library(name: "LuxAnalytics", targets: ["LuxAnalytics"])
     ],
     targets: [
-        .target(name: "LuxAnalytics", dependencies: []),
+        .target(
+            name: "LuxAnalytics",
+            dependencies: [],
+            resources: [
+                .process("PrivacyInfo.xcprivacy")
+            ],
+            swiftSettings: [
+                // We want Swift 6 with all its safety features!
+            ]
+        ),
         .testTarget(name: "LuxAnalyticsTests", dependencies: ["LuxAnalytics"])
     ]
 )
