@@ -1,6 +1,7 @@
 import Foundation
 
 public struct AnalyticsEvent: Codable, Sendable {
+    public let id: String
     public let name: String
     public let timestamp: String
     public let userId: String?
@@ -8,6 +9,7 @@ public struct AnalyticsEvent: Codable, Sendable {
     public let metadata: [String: String]
     
     public init(name: String, timestamp: String, userId: String?, sessionId: String?, metadata: [String: String]) {
+        self.id = UUID().uuidString
         self.name = name
         self.timestamp = timestamp
         self.userId = userId
@@ -16,8 +18,15 @@ public struct AnalyticsEvent: Codable, Sendable {
     }
     
     enum CodingKeys: String, CodingKey {
-        case name, timestamp, metadata
+        case id, name, timestamp, metadata
         case userId = "user_id"
         case sessionId = "session_id"
+    }
+    
+    func toDictionary() throws -> [String: Any] {
+        let encoder = JSONEncoder()
+        let data = try encoder.encode(self)
+        let json = try JSONSerialization.jsonObject(with: data)
+        return json as! [String: Any]
     }
 }
